+++ temba/api/v2/tests/test_ticket_actions.py
@@ -125,12 +125,15 @@ class TicketActionsEndpointTest(APITest):
         self.assertEqual(sales, ticket2.topic)
 
         # close tickets
-        self.assertPost(
+        response = self.assertPost(
             endpoint_url,
             self.agent,
             {"tickets": [str(ticket1.uuid), str(ticket2.uuid)], "action": "close"},
-            status=204,
+            status=200,
         )
+
+        # ticket1 reported as failure because it is already closed
+        self.assertEqual({"failures": [str(ticket1.uuid)]}, response.json())
 
         ticket1.refresh_from_db()
         ticket2.refresh_from_db()